Differ.merge -> mergeAllBranches (#61)
this PR does two main things:
1. Change differ.merge to manage merging all head branches
2. Combine all commit metadata into one field, and change where it’s set 

It has a number of breaking changes:
1. need to implement `differ.mergeAllBranches` (see below)
2. `differ.computeRef` no longer takes metadata (this means the ref computation will be different!)
3. `differ.migrate` now is expected to return a `{ doc, metadata }` structure (instead of `{ doc, editMetadata }`

## Differ.mergeAllBranches

Instead of `mergeHeads` being internal to `TrimergeClient`, `Differ` now provides a `mergeAllBranches` function. It is expected to try to merge all head refs on the graph.

This function (see differ.ts) is handed a `branchHeadRefs: string[]` and set of helper functions for getting commit graph info, docs, and adding individual merge commits.

To mimic the existing `mergeHeads` behavior, a `makeAutoBranchMerger` is provided. It takes a `SortRefsFn`, and `MergeDocFn` (almost identical to the existing `differ.merge` type, except that it renames `editMetadata` to `metadata`). (To fully reproduce the existing behavior, sortRefs can be defined as a string sort `(refA: string, refB: string) => refA < refB ? -1 : 1` — but something that introspects the metadata is possibly more appropriate)

## New commit metadata and behavior

The main change here is that `metadata` can contain data from various sources (app, client, local store, remote) and grouped together. The metadata is now all accessible when doing merges.

- `CommitAck` (in `AckCommitsEvent`) now can specify new metadata that is used to overwrite the existing metadata
- `editMetadata` renamed to `metadata` in various places
- commits no longer have `userId`, `clientId` embedded in them
- notion of `main` is removed (can be added in user-land)
- TrimergeClient now takes an optional `adddNewCommitMetadata(metadata, ref, userId, clientId)` that is called to add metadata (like a timestamp, clientId) to new commits (via migrate, edit, and merge)
- `trimerge-indexed-db`:
  - now takes an optional `addStoreMetadata(commit, localStoreId, commitIndex)` which can be used to add metadata before a commit is stored locally (and is sent back in the ack)
  - if the remote adds a commit we already have, replace local metadata
  - if the remote acks a commit with new metadata, replace local metadata